requirements.txt
flask
flask-cors
flask-socketio
requests
beautifulsoup4
transformers
torch
python-dotenv
eventlet